Separate dimension controls, update progress
This makes it possible to click (& optionally drag) the progress bar as an alternative way to change the slice or time point.
